python_backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
aptos-sdk>=0.1.0
python-dotenv>=0.19.0 